+++ src/startServer.ts
@@ -1,14 +1,13 @@
-import { WasmUrls } from '@cucumber/language-service'
 import { TextDocuments } from 'vscode-languageserver'
 import { createConnection, ProposedFeatures } from 'vscode-languageserver/node'
 import { TextDocument } from 'vscode-languageserver-textdocument'
 
 import { CucumberLanguageServer } from './CucumberLanguageServer.js'
 
-export function startServer(wasmUrls: WasmUrls) {
+export function startServer() {
   const connection = createConnection(ProposedFeatures.all)
   const documents = new TextDocuments(TextDocument)
-  new CucumberLanguageServer(connection, documents, wasmUrls)
+  new CucumberLanguageServer(connection, documents)
   connection.listen()
 
   // Don't die on unhandled Promise rejections