+++ test/CucumberLanguageServer.test.ts
@@ -1,4 +1,3 @@
-import { WasmUrls } from '@cucumber/language-service'
 import assert from 'assert'
 import { Duplex } from 'stream'
 import { Logger, StreamMessageReader, StreamMessageWriter } from 'vscode-jsonrpc/node'
@@ -25,11 +24,6 @@ import { TextDocument } from 'vscode-languageserver-textdocument'
 
 import { CucumberLanguageServer } from '../src/CucumberLanguageServer.js'
 import { Settings } from '../src/types'
-
-const wasmUrls: WasmUrls = {
-  java: `node_modules/@cucumber/language-service/tree-sitter-java.wasm`,
-  typescript: `node_modules/@cucumber/language-service/tree-sitter-typescript.wasm`,
-}
 
 describe('CucumberLanguageServer', () => {
   let inputStream: Duplex
@@ -59,7 +53,7 @@ describe('CucumberLanguageServer', () => {
     serverConnection = createConnection(inputStream, outputStream)
     documents = new TextDocuments(TextDocument)
 
-    new CucumberLanguageServer(serverConnection, documents, wasmUrls)
+    new CucumberLanguageServer(serverConnection, documents)
     serverConnection.listen()
 
     const initializeParams: InitializeParams = {